С помощью вертикального наследования перенес фикстуру в TestBase, добавил основу для управления браузером
